requirements.txt
Django==4.2.7
django-markdownx
django-embed-video
Pillow
django-crispy-forms
crispy-bootstrap5
python-decouple
setuptools
djangorestframework
django-cors-headers
celery
redis
numpy
torch
transformers
scikit-learn
pandas